+++ frontend/providers/devbox/message/en.json
@@ -33,6 +33,8 @@
   "basic_configuration": "Basic",
   "basic_info": "Basic",
   "cancel": "Cancel",
+  "choose_a_image_to_update": "Choose a image to update",
+  "choose_the_image_to_update": "Choose the image to update",
   "code_server": "CodeServer",
   "config_form": "Form",
   "confirm": "Confirm",
@@ -41,6 +43,7 @@
   "confirm_update_devbox": "Changing the CPU or memory will cause the project to restart and any active IDE instances will need to reconnect. Are you sure you wish to continue?",
   "control": "action",
   "convert_to_runtime": "Convert to Template",
+  "convert_to_template": "Convert to template",
   "copy": "Copy",
   "copy_failed": "Copy failed",
   "copy_success": "Copy succeeded",
@@ -52,6 +55,7 @@
   "create_devbox_first": "Create your first Devbox",
   "create_directly": "Want to deploy a new app directly?",
   "create_failed": "Creation failed",
+  "create_new": "Create new",
   "create_or_update_template": "Create or Update Template",
   "create_success": "Creation succeeded",
   "create_template": "Create Template",
@@ -229,10 +233,10 @@
   "refresh": "Refresh",
   "release": "Release",
   "release_confirm_info": "Your Devbox will be temporarily shut down during the release process. Please ensure all changes are saved before proceeding.",
-  "release_failed": "Release failed",
+  "release_failed": "Failed",
   "release_pending": "Releasing...",
   "release_prompt": "The release process will shutdown your DevBox for processing, once complete your DevBox will return to its current state.",
-  "release_success": "Release success",
+  "release_success": "success",
   "release_successful": "Release succeeded",
   "release_version": "Release",
   "remaining": "Available",
@@ -280,6 +284,7 @@
   "template_repository_name_already_exists": "This name is already taken. Please try another one.",
   "template_search": "Search",
   "template_ssaved_successfully": "Template saved successfully!",
+  "template_update": "Update",
   "terminal": "Terminal",
   "to_update": "Update",
   "total": "Total",
@@ -288,6 +293,7 @@
   "update": "Update",
   "update Time": "Updated At",
   "update_app": "update/create",
+  "update_current": "Update current",
   "update_devbox": "Update DevBox",
   "update_failed": "Update failed",
   "update_matched_apps_notes": "Or, you can update existing applications",
