+++ frontend/providers/devbox/message/zh.json
@@ -33,6 +33,8 @@
   "basic_configuration": "基础配置",
   "basic_info": "基础信息",
   "cancel": "取消",
+  "choose_a_image_to_update": "选择一个镜像更新",
+  "choose_the_image_to_update": "选择镜像去更新",
   "code_server": "CodeServer",
   "config_form": "配置表单",
   "confirm": "确认",
@@ -42,6 +44,7 @@
   "continue": "继续",
   "control": "操作",
   "convert_to_runtime": "转换成模板",
+  "convert_to_template": "转换成模板",
   "copy": "复制",
   "copy_failed": "复制失败",
   "copy_success": "复制成功",
@@ -53,6 +56,7 @@
   "create_devbox_first": "创建你的第一个 Devbox",
   "create_directly": "想要直接上线新应用？",
   "create_failed": "创建失败",
+  "create_new": "创建新模板",
   "create_or_update_template": "创建/更新模板",
   "create_success": "创建成功",
   "create_template": "创建模板",
@@ -282,6 +286,7 @@
   "template_repository_name_already_exists": "输入的名称已存在，请重新输入。",
   "template_search": "搜索模板",
   "template_ssaved_successfully": "模板保存成功！",
+  "template_update": "更新",
   "terminal": "终端",
   "to_update": "去更新",
   "total": "总共",
@@ -290,6 +295,7 @@
   "update": "变更",
   "update Time": "更新时间",
   "update_app": "更新应用镜像/上线新应用",
+  "update_current": "更新现有模板",
   "update_devbox": "变更项目",
   "update_failed": "变更失败",
   "update_matched_apps_notes": "或者你可以更新已有应用：",
